+++ packages/util-crypto/src/blake2/asU8a.ts
@@ -1,4 +1,4 @@
-// Copyright 2017-2024 @polkadot/util-crypto authors & contributors
+// Copyright 2017-2025 @polkadot/util-crypto authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import { blake2b as blake2bJs } from '@noble/hashes/blake2b';