+++ packages/util-crypto/src/hd/ethereum/index.spec.ts
@@ -1,4 +1,4 @@
-// Copyright 2017-2024 @polkadot/util-crypto authors & contributors
+// Copyright 2017-2025 @polkadot/util-crypto authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 /// <reference types="@polkadot/dev-test/globals.d.ts" />
@@ -39,14 +39,14 @@ describe('hdEthereum', (): void => {
   ]);
 
   it('derives the right key pair from a mnemonic', (): void => {
-    const key = hdEthereum(mnemonicToLegacySeed(PHRASE, '', false, 64));
+    const key = hdEthereum(mnemonicToLegacySeed(PHRASE, '', false, 64, 2048));
 
     expect(key.publicKey).toEqual(PUBLIC);
     expect(key.secretKey).toEqual(SECRET);
   });
 
   it('derives the right key pair from a mnemonic and a derivation path', (): void => {
-    const key = hdEthereum(mnemonicToLegacySeed(PHRASE, '', false, 64), derivationPath);
+    const key = hdEthereum(mnemonicToLegacySeed(PHRASE, '', false, 64, 2048), derivationPath);
 
     expect(key.publicKey).toEqual(PUBLICDERIVED);
     expect(key.secretKey).toEqual(SECRETDERIVED);